DXSysStats: Update to latest trunk version + work towards compilation
--HG--
branch : DesktopX
extra : convert_revision : svn%3Aab80709b-eb45-0410-bb3a-633ce738720d/trunk/DesktopX%40327

+++ Plugins/DXSysStats/COMOverlays/ScrollerOverlayDialog.cpp
@@ -50,7 +50,7 @@ LRESULT ScrollerOverlayDialog::alpha(WORD wNotifyCode, WORD wID, HWND hWndCtl, B
 
 	if (iRet)
 	{
-		((CScrollerOverlay*)overlay)->put_Alpha(i);
+		((CScrollerOverlay*)overlay)->put_Alpha((BYTE)i);
 		notifyObserver();
 	}
